Refactor und erweitere die Struktur der Anwendung mit neuen Bildschirmen und Modellen
- Aktualisiere die Importpfade in router.dart, um die neue Struktur widerzuspiegeln.
- Füge neue Dateien für App-Text, Snackbar-Helfer, Spieltypen, Spieler- und Wiederverbindungsdaten hinzu.
- Implementiere LobbyService und ReconnectService zur Verwaltung von Lobbys und Wiederverbindungen.
- Erstelle neue Bildschirme für die Spielauswahl, Lobbyerstellung, Lobbyansicht und Spielansicht.
- Entferne veraltete Dateien, um die Codebasis zu bereinigen und die Wartbarkeit zu verbessern.
