+++ lib/data/services/device_id_helper.dart
@@ -1,4 +1,4 @@
-// Datei: lib/helpers/device_id_helper.dart
+// Datei: lib/data/services/device_id_helper.dart
 
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:uuid/uuid.dart';
