+++ lib/presentation/screens/game_selection_screen.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:jugend_app/model/game_type.dart';
+import 'package:jugend_app/data/models/game_type.dart';
 
 class GameSelectionScreen extends StatelessWidget {
   const GameSelectionScreen({super.key});
