+++ lib/presentation/screens/lobby_join_screen.dart
@@ -2,10 +2,10 @@
 
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:jugend_app/model/reconnect_data.dart';
-import 'package:jugend_app/services/lobby_service.dart';
-import 'package:jugend_app/services/reconnect_service.dart';
-import 'package:jugend_app/helpers/snackbar_helper.dart';
+import 'package:jugend_app/data/models/reconnect_data.dart';
+import 'package:jugend_app/data/services/lobby_service.dart';
+import 'package:jugend_app/data/services/reconnect_service.dart';
+import 'package:jugend_app/core/snackbar_helper.dart';
 
 class LobbyJoinScreen extends StatefulWidget {
   const LobbyJoinScreen({super.key});
